add decimals.js and remove alternativeContract call for getDecimals
