+++ src/stores/utils/web3.js
@@ -1,6 +1,5 @@
 import Web3 from 'web3'
 import Web3Utils from 'web3-utils'
-import BN from 'bignumber.js'
 
 const getWeb3 = () => {
   return new Promise(function (resolve, reject) {
@@ -76,42 +75,6 @@ export const estimateGas = async (web3, to, gasPrice, from, value, data) =>{
   const gas = await web3.eth.estimateGas({to, gasPrice, from, value, data})
   return Web3Utils.toHex(gas.toString())
 }
-
-export const fromDecimals = (number, decimals) => {
-  if (decimals == null) {
-    decimals = 18;
-  }
-  const returnValue = toBigNumber(number).dividedBy(Math.pow(10, decimals));
-  return isBigNumber(number) ? returnValue : returnValue.toString(10);
-};
-
-export const toDecimals = (number, decimals) => {
-  if (decimals == null) {
-    decimals = 18;
-  }
-  const returnValue = toBigNumber(number).times(Math.pow(10, decimals));
-  return isBigNumber(number) ? returnValue : returnValue.toString(10);
-};
-
-const isBigNumber = (object) => {
-    return (object && (object instanceof BN || (object.constructor && object.constructor.name === 'BigNumber')));
-};
-
-const toBigNumber = (number) => {
-    /*jshint maxcomplexity:5 */
-    number = number || 0;
-    if (isBigNumber(number))
-        return number;
-    if (isString(number) && (number.indexOf('0x') === 0 || number.indexOf('-0x') === 0)) {
-        return new BN(number.replace('0x',''), 16);
-    }
-    return new BN(number.toString(10), 10);
-};
-
-const isString = (object) => {
-    return typeof object === 'string' ||
-        (object && object.constructor && object.constructor.name === 'String');
-};
 
 const processWeb3 = (web3, resolve,  reject) => {
   web3.version.getNetwork((err, netId) => {
